Done create modules create backend

+++ app/Http/Controllers/Admin/ModulsController.php
@@ -44,6 +44,7 @@ class ModulsController extends Controller{
 
         //CREATE FILE
         if($request->type === 'Livewire'){
+          
             $folderName = $request->controller;
             $contentName = $request->controller.'Content';
             $viewFolderName = preg_split('/(?=[A-Z])/',$request->controller);
@@ -51,13 +52,16 @@ class ModulsController extends Controller{
 
             $finalViewFolderName = strtolower($viewFolderName[1])."-".strtolower($viewFolderName[2]);
             $finalViewContentName = strtolower($viewContentName[1])."-".strtolower($viewContentName[2]).'-'.'content';
-      
+
+            if(file_exists(base_path('app/Livewire/Component/ModuleContents/'.$folderName.'/'.$contentName.'.php'))){
+               return CommonHelpers::redirect(CommonHelpers::mainpath(), "Content already exist", 'error');
+            }
             //MAKE FOLDER
             $folder = base_path('app/Livewire/Component/ModuleContents/'.$folderName);
             File::makeDirectory($folder, $mode = 0777, true, true);
             //MAKE FILE CONTENT
             $path = base_path("app/Livewire/Component/ModuleContents/$folderName/");
-            $php = self::livewireContent($contentName,$folderName);
+            $php = self::livewireContent($contentName,$folderName,$finalViewFolderName,$finalViewContentName);
             $php = trim($php);
             file_put_contents($path.$contentName.'.php', $php);
             //MAKE FOLDER VIEW CONTENT
@@ -65,44 +69,69 @@ class ModulsController extends Controller{
             File::makeDirectory($makeFolderViewContent, $mode = 0777, true, true);
             //MAKE VIEW CONTENT
             $pathView = base_path("resources/views/livewire/component/module-contents/$finalViewFolderName/");
-            $viewContent = self::viewContent();
+            $viewContent = self::viewContentLivewire();
             $viewContent = trim($viewContent);
             file_put_contents($pathView.$finalViewContentName.'.blade.php', $viewContent);
-            //CREATE MODULE
-            DB::table('ad_modules')->insert([
-                'name'         => $request->name,
-                'icon'         => $request->icon,
-                'path'         => $request->path,
-                'table_name'   => NULL,
-                'controller'   => $folderName."\\".$contentName,
-                'is_protected' => 0,
-                'is_active'    => 1,
-                'created_at'   => date('Y-m-d H:i:s')
-            ]);
-            //CREATE MENUS
-            $menusId = DB::table('ad_menuses')->insertGetId([
-                'name'                => $request->name,
-                'type'                => 'Route',
-                'icon'                => $request->icon,
-                'path'                => $folderName."\\".$contentName.'GetIndex',
-                'color'               => NULL,
-                'parent_id'           => 0,
-                'is_active'           => 1,
-                'is_dashboard'        => 0,
-                'id_ad_privileges'    => 1,
-                'sorting'             => 1,
-                'created_at'          => date('Y-m-d H:i:s')
-            ]);
-            //CREATE MENUS PRIVILEGE
-            DB::table('ad_menus_privileges')->insert(['id_ad_menus' => $menusId, 'id_ad_privileges' => CommonHelpers::myPrivilegeId()]);
 
-        }else{
+            //CREATE VIEW MODULE CONTENT LIVEWIRE
+            //MAKE FOLDER VIEW CONTENT MODULE
+            $makeFolderViewContentModule = base_path('resources/views/modules/'.$finalViewFolderName);
+            File::makeDirectory($makeFolderViewContentModule, $mode = 0777, true, true);
+            $pathViewModule = base_path("resources/views/modules/$finalViewFolderName/");
+            $viewContentModule = self::viewContent();
+            $viewContentModule = trim($viewContentModule);
+            file_put_contents($pathViewModule.$finalViewContentName.'.blade.php', $viewContentModule);
+
+            //CREATE MODULE
+            DB::table('ad_modules')->updateOrInsert([
+                    'name'         => $request->name,
+                    'path'         => $request->path,
+                    'controller'   => $folderName."\\".$contentName
+                ],
+                [
+                    'name'         => $request->name,
+                    'icon'         => $request->icon,
+                    'path'         => $request->path,
+                    'table_name'   => NULL,
+                    'controller'   => $folderName."\\".$contentName,
+                    'is_protected' => 0,
+                    'is_active'    => 1,
+                    'created_at'   => date('Y-m-d H:i:s')
+                ]
+            );
+            //CREATE MENUS
+            $isExist = DB::table('ad_menuses')->where('name',$request->name)->where('path',$folderName."\\".$contentName.'GetIndex')->exists();
+            if(!$isExist){
+                $menusId = DB::table('ad_menuses')->insertGetId(
+                    [
+                        'name'                => $request->name,
+                        'type'                => 'Route',
+                        'icon'                => $request->icon,
+                        'path'                => $folderName."\\".$contentName.'GetIndex',
+                        'color'               => NULL,
+                        'parent_id'           => 0,
+                        'is_active'           => 1,
+                        'is_dashboard'        => 0,
+                        'id_ad_privileges'    => 1,
+                        'sorting'             => 0,
+                        'created_at'          => date('Y-m-d H:i:s')
+                    ]
+                );
+                //CREATE MENUS PRIVILEGE
+                DB::table('ad_menus_privileges')->insert(['id_ad_menus' => $menusId, 'id_ad_privileges' => CommonHelpers::myPrivilegeId()]);
+            }
+        }else if($request->type === 'Admin Controller'){
             $controllerName = $request->controller.'Controller';
             $viewFileName = preg_split('/(?=[A-Z])/',$request->controller);
             $finalViewFileName = strtolower($viewFileName[1])."-".strtolower($viewFileName[2]);
+
+            if(file_exists(base_path('app/Http/Controllers/Admin/'.$controllerName.'.php'))){
+              return CommonHelpers::redirect(CommonHelpers::mainpath(), "Controller already exist", 'error');
+            }
+
             //MAKE FILE CONTROLLER
             $path = base_path("app/Http/Controllers/Admin/");
-            $php = self::controllerContent($controllerName, $finalViewFileName);
+            $php = self::adminControllerContent($controllerName, $finalViewFileName);
             $php = trim($php);
             file_put_contents($path.$controllerName.'.php', $php);
             //MAKE FILE VIEW
@@ -114,54 +143,130 @@ class ModulsController extends Controller{
             $viewContent = trim($viewContent);
             file_put_contents($pathView.$finalViewFileName.'.blade.php', $viewContent);
               //CREATE MODULE
-              DB::table('ad_modules')->insert([
-                'name'         => $request->name,
-                'icon'         => $request->icon,
-                'path'         => $request->path,
-                'table_name'   => NULL,
-                'controller'   => $controllerName,
-                'is_protected' => 0,
-                'is_active'    => 1,
-                'created_at'   => date('Y-m-d H:i:s')
-            ]);
+              DB::table('ad_modules')->updateOrInsert([
+                    'name'         => $request->name,
+                    'path'         => $request->path,
+                    'controller'   => $controllerName
+                ],
+                [
+                    'name'         => $request->name,
+                    'icon'         => $request->icon,
+                    'path'         => $request->path,
+                    'table_name'   => NULL,
+                    'controller'   => $controllerName,
+                    'is_protected' => 0,
+                    'is_active'    => 1,
+                    'created_at'   => date('Y-m-d H:i:s')
+                ]
+            );
             //CREATE MENUS
-            $menusId = DB::table('ad_menuses')->insertGetId([
-                'name'                => $request->name,
-                'type'                => 'Route',
-                'icon'                => $request->icon,
-                'path'                => $controllerName.'GetIndex',
-                'color'               => NULL,
-                'parent_id'           => 0,
-                'is_active'           => 1,
-                'is_dashboard'        => 0,
-                'id_ad_privileges'    => 1,
-                'sorting'             => 1,
-                'created_at'          => date('Y-m-d H:i:s')
-            ]);
-            //CREATE MENUS PRIVILEGE
-            DB::table('ad_menus_privileges')->insert(['id_ad_menus' => $menusId, 'id_ad_privileges' => CommonHelpers::myPrivilegeId()]);
+            $isExist = DB::table('ad_menuses')->where('name',$request->name)->where('path',$controllerName.'GetIndex')->exists();
+            if(!$isExist){
+                $menusId = DB::table('ad_menuses')->insertGetId(
+                    [
+                        'name'                => $request->name,
+                        'type'                => 'Route',
+                        'icon'                => $request->icon,
+                        'path'                => $controllerName.'GetIndex',
+                        'color'               => NULL,
+                        'parent_id'           => 0,
+                        'is_active'           => 1,
+                        'is_dashboard'        => 0,
+                        'id_ad_privileges'    => 1,
+                        'sorting'             => 0,
+                        'created_at'          => date('Y-m-d H:i:s')
+                    ]
+                );
+                //CREATE MENUS PRIVILEGE
+                DB::table('ad_menus_privileges')->insert(['id_ad_menus' => $menusId, 'id_ad_privileges' => CommonHelpers::myPrivilegeId()]);
+            }
+        }else{
+            $folderName = $request->controller;
+            $contentName = $request->controller.'Controller';
+            $viewFolderName = preg_split('/(?=[A-Z])/',$request->controller);
+            $viewContentName = preg_split('/(?=[A-Z])/',$request->controller);
 
+            $finalViewFolderName = strtolower($viewFolderName[1])."-".strtolower($viewFolderName[2]);
+            $finalViewContentName = strtolower($viewContentName[1])."-".strtolower($viewContentName[2]).'-'.'content';
+
+            if(file_exists(base_path('app/Http/Controllers/'.$folderName.'/'.$contentName.'.php'))){
+               return CommonHelpers::redirect(CommonHelpers::mainpath(), "Controller already exist", 'danger');
+            }else{
+                //MAKE FOLDER
+                $folder = base_path('app/Http/Controllers/'.$folderName);
+                File::makeDirectory($folder, $mode = 0777, true, true);
+                //MAKE FILE CONTENT
+                $path = base_path("app/Http/Controllers/$folderName/");
+                $php = self::controllerContent($contentName,$folderName,$finalViewFolderName,$finalViewContentName);
+                $php = trim($php);
+                file_put_contents($path.$contentName.'.php', $php);
+                //MAKE FOLDER VIEW CONTENT
+                $makeFolderViewContent = base_path('resources/views/'.$finalViewFolderName);
+                File::makeDirectory($makeFolderViewContent, $mode = 0777, true, true);
+            
+                //MAKE FILE CONTROLLER
+                $pathViewController = base_path("resources/views/$finalViewFolderName/");
+                $viewContent = self::viewContent();
+                $viewContent = trim($viewContent);
+                file_put_contents($pathViewController.$finalViewContentName.'.blade.php', $viewContent);
+
+                //CREATE MODULE
+                DB::table('ad_modules')->updateOrInsert([
+                        'name'         => $request->name,
+                        'path'         => $request->path,
+                        'controller'   => $folderName."\\".$contentName
+                    ],
+                    [
+                        'name'         => $request->name,
+                        'icon'         => $request->icon,
+                        'path'         => $request->path,
+                        'table_name'   => NULL,
+                        'controller'   => $folderName."\\".$contentName,
+                        'is_protected' => 0,
+                        'is_active'    => 1,
+                        'created_at'   => date('Y-m-d H:i:s')
+                    ]
+                );
+                //CREATE MENUS
+                $isExist = DB::table('ad_menuses')->where('name',$request->name)->where('path',$folderName."\\".$contentName.'GetIndex')->exists();
+                if(!$isExist){
+                    $menusId = DB::table('ad_menuses')->insertGetId([
+                        'name'                => $request->name,
+                        'type'                => 'Route',
+                        'icon'                => $request->icon,
+                        'path'                => $folderName."\\".$contentName.'GetIndex',
+                        'color'               => NULL,
+                        'parent_id'           => 0,
+                        'is_active'           => 1,
+                        'is_dashboard'        => 0,
+                        'id_ad_privileges'    => 1,
+                        'sorting'             => 0,
+                        'created_at'          => date('Y-m-d H:i:s')
+                    ]);
+                    //CREATE MENUS PRIVILEGE
+                    DB::table('ad_menus_privileges')->insert(['id_ad_menus' => $menusId, 'id_ad_privileges' => CommonHelpers::myPrivilegeId()]);
+                }
+            
+                CommonHelpers::redirect(CommonHelpers::mainpath(), "Modules created successfully", 'success');
+            } 
         }
-
-        CommonHelpers::redirect(CommonHelpers::mainpath(), "Modules created successfully", 'success');
-    
     }
 
-    public function livewireContent($contentName, $folderName){
+    public function livewireContent($contentName, $folderName,$finalViewFolderName,$finalViewContentName){
         return '<?php
-        namespace App\Livewire\Component\ModuleContents\"'.$folderName'";
+        namespace App\Livewire\Component\ModuleContents\"'.$folderName.'";
         use Livewire\Component;
         use App\Helpers\CommonHelpers;
         
         class '.$contentName.' extends Component{
             public function index(){}
             public function render(){
-                return view("livewire.component.module-contents.employee-accounts.employee-accounts-content");
+                return view("livewire.component.module-contents.'.$finalViewFolderName.'.'.$finalViewContentName.'");
             }
         } ?>';
     }
 
-    public function controllerContent($controllerName, $finalViewFileName){
+    public function adminControllerContent($controllerName, $finalViewFileName){
         return '<?php
 
                 namespace App\Http\Controllers\Admin; 
@@ -181,9 +286,34 @@ class ModulsController extends Controller{
                 ?>';
     }
 
+    public function controllerContent($controllerName, $finalViewFileName,$finalViewFolderName,$finalViewContentName){
+        return '<?php
+
+                namespace App\Http\Controllers\"'.$finalViewFileName.'"; 
+                use App\Helpers\CommonHelpers;
+                use App\Http\Controllers\Controller;
+                use Illuminate\Http\Request;
+                use Illuminate\Http\RedirectResponse;
+                use Illuminate\Support\Facades\Auth;
+                use Illuminate\Support\Facades\Session;
+                use DB;
+        
+                class '.$controllerName.' extends Controller{
+                    public function getIndex(){
+                        return view("'.$finalViewFolderName.'/'.$finalViewContentName.'");
+                    }
+                }
+                ?>';
+    }
+
+    public function viewContentLivewire(){
+        return '<div> This is livewire view content</div>';
+    }
+
     public function viewContent(){
-        return '@extend("layout")
+        return '@extends("layout")
                 @section("content")
+                Extend livewire here if content is livewire...
                 @endsection';
     }
